Updated INDEX.EJS
Updated INDEX.EJS file.
Added more semantic ui elements.
Formatted the date string to a more readable format.
Added link to show a specific item.
Truncated the description in the body so all of it is not showing in the
index page.  The user will have to click on the link to view the whole
description.